version 1.0.0 release

+++ back/src/post/post.service.ts
@@ -1,39 +1,77 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { PAGE_LIMIT, UPLOAD_LIMIT } from 'common/constants/nums';
 import { Comment } from 'src/comments/entities/comment.entity';
 import { CreateContentDto } from 'src/contents/dto/create-content.dto';
 import { Content } from 'src/contents/entities/content.entity';
 import { PostContent } from 'src/post-contents/entities/post-content.entity';
+import { User } from 'src/users/entities/user.entity';
+import { UserRepository } from 'src/users/user.repository';
 import { Connection } from 'typeorm';
+import { convertStringToNumber } from 'utils/value-converter.util';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PatchPostRequestDto } from './dto/patchPostRequestDto';
 import { Post } from './entities/post.entity';
 import { PostRepository } from './post.repository';
-
-const LIMIT_NUMBER = 10;
-const USER_ID = 1;
 
 @Injectable()
 export class PostService {
   constructor(
     @InjectRepository(PostRepository)
     private postRepository: PostRepository,
+    private userRepository: UserRepository,
     private connection: Connection
   ) {}
 
   async create(createPostDto: CreatePostDto, contentsInfos: CreateContentDto[], userId: number) {
-    return await this.postRepository.createPost(createPostDto, contentsInfos, userId);
+    const user = await this.userRepository.findOne(userId, { relations: ['species'] });
+
+    if (!user)
+      throw new HttpException('Error: 존재하지 않는 사용자입니다.', HttpStatus.BAD_REQUEST);
+
+    const animalContent = translateHumanToAnimal(user, createPostDto);
+
+    return await this.postRepository.createPost(
+      createPostDto,
+      contentsInfos,
+      userId,
+      animalContent
+    );
   }
 
-  async findAll(habitatId: number, lastPostId?: number) {
+  async findAll(habitatId: number, user: any, lastPostId?: number) {
+    const userId = user ? user.userId : user;
+
     if (!lastPostId) {
-      return await this.getFirstPage(habitatId, USER_ID);
+      const result = await this.getFirstPage(habitatId, userId);
+
+      result.posts.forEach((post) =>
+        convertStringToNumber(post, 'numOfHearts', 'numOfComments', 'is_heart')
+      );
+
+      return result;
     } else {
-      return await this.getNextPage(habitatId, USER_ID, lastPostId);
+      const result = await this.getNextPage(habitatId, userId, lastPostId);
+
+      result.posts.forEach((post) =>
+        convertStringToNumber(post, 'numOfHearts', 'numOfComments', 'is_heart')
+      );
+
+      return result;
     }
   }
 
-  async getFirstPage(habitatId: number, userId: number) {
+  async findAllByUserId(userId: number, lastPostId?: number): Promise<UserPostDto[]> {
+    const results = await this.postRepository.findAllByUserId(userId, lastPostId);
+
+    results.forEach((result: UserPostDto) =>
+      convertStringToNumber(result, 'numOfHearts', 'numOfComments')
+    );
+
+    return results;
+  }
+
+  async getFirstPage(habitatId: number, userId: number | false) {
     let baseSql = this.getBaseQuery();
     let tailSql = this.getTailQuery();
     let whereSql = `where p.habitat_id = ? 
@@ -42,16 +80,13 @@ export class PostService {
     const posts = await this.connection.query(baseSql + whereSql + tailSql, [
       userId,
       habitatId,
-      LIMIT_NUMBER,
+      PAGE_LIMIT,
     ]);
 
-    const lastPostId = posts.length === LIMIT_NUMBER ? posts[LIMIT_NUMBER - 1].post_id : null;
-
-    if (lastPostId) return { posts, lastPostId };
-    else return { posts };
+    return { posts };
   }
 
-  async getNextPage(habitatId: number, userId: number, oldLastPostId: number) {
+  async getNextPage(habitatId: number, userId: number | false, lastPostId: number) {
     let baseSql = this.getBaseQuery();
     let tailSql = this.getTailQuery();
     let whereSql = `where p.habitat_id = ?
@@ -63,29 +98,25 @@ export class PostService {
     const posts = await this.connection.query(baseSql + whereSql + middleSql + tailSql, [
       userId,
       habitatId,
-      oldLastPostId,
-      LIMIT_NUMBER,
+      lastPostId,
+      PAGE_LIMIT,
     ]);
 
-    const currentLastPostId =
-      posts.length === LIMIT_NUMBER ? posts[LIMIT_NUMBER - 1].post_id : null;
-
-    if (currentLastPostId) return { posts, lastPostId: currentLastPostId };
-    else return { posts };
+    return { posts };
   }
 
   private getBaseQuery() {
     return `
-    select p.post_id, p.human_content, p.animal_content, p.created_at, u.user_id, u.username, u.nickname, c.url as user_image_url
-    , group_concat(distinct pcc.url) as post_contents_urls, group_concat(distinct pcc.mime_type) as post_contents_types
+    select p.post_id, p.human_content, p.animal_content, p.created_at, u.user_id, u.username, u.nickname, c.contents_id, c.url as user_image_url
+    , group_concat(pcc.url order by pcc.contents_id) as post_contents_urls, group_concat(pcc.contents_id order by pcc.contents_id) as post_contents_ids
     , (select count(*) from heart where post_id = p.post_id) as numOfHearts
     , (select count(*) from comment where post_id = p.post_id) as numOfComments
-    , if((select count(*) from heart where post_id = p.post_id and user_id = ?), true, false) as is_heart
+    , (select count(*) from heart where post_id = p.post_id and user_id = ?) as is_heart
     from post p
-    inner join user u on u.user_id = p.user_id
-    inner join contents c on c.contents_id = u.contents_id
-    inner join post_contents pc on pc.post_id = p.post_id
-    inner join contents pcc on pc.contents_id = pcc.contents_id
+    straight_join user u on u.user_id = p.user_id
+    left join contents c on c.contents_id = u.contents_id
+    straight_join post_contents pc on pc.post_id = p.post_id
+    straight_join contents pcc on pc.contents_id = pcc.contents_id
     `;
   }
 
@@ -97,65 +128,80 @@ export class PostService {
     `;
   }
 
-  async findOne(id: number) {
+  async findOne(posId: number, user: any) {
+    const userId = user ? user.userId : user;
     let baseSql = this.getBaseQuery();
     let whereSql = `where p.post_id = ?;
     `;
-    return (await this.connection.query(baseSql + whereSql, [USER_ID, id]))[0];
+
+    const result = (await this.connection.query(baseSql + whereSql, [userId, posId]))[0];
+
+    if (!result) throw new HttpException('Error: 잘못된 접근입니다.', HttpStatus.BAD_REQUEST);
+
+    convertStringToNumber(result, 'numOfHearts', 'numOfComments', 'is_heart');
+    return result;
   }
 
   async update(
-    id: number,
+    postId: number,
     patchPostRequestDto: PatchPostRequestDto,
-    contentsInfos: CreateContentDto[]
+    contentsInfos: CreateContentDto[],
+    userId: number
   ) {
     const queryRunner = this.connection.createQueryRunner();
     await queryRunner.connect();
 
+    const userRepository = queryRunner.manager.getRepository(User);
     const postRepository = queryRunner.manager.getRepository(Post);
     const postContentRepository = queryRunner.manager.getRepository(PostContent);
     const contentRepository = queryRunner.manager.getRepository(Content);
 
-    const post = await postRepository.findOne(id, {
+    const post = await postRepository.findOne(postId, {
       relations: ['postContents'],
     });
 
-    const updateIds = JSON.parse(patchPostRequestDto.contentIds).map(
-      (info: { id: string }, i: number) => info.id
-    );
+    const updateIds = patchPostRequestDto.contentIds.split(',').map((id: string) => id);
 
     const excludedPostContents = post.postContents.filter(
-      (postContent, i) => !updateIds.includes('' + postContent.contentsId)
+      (postContent) => !updateIds.includes('' + postContent.contentsId)
     );
 
     if (
       post.postContents.length - excludedPostContents.length + contentsInfos.length >
-      LIMIT_NUMBER
+      UPLOAD_LIMIT
     )
       return false;
 
-    queryRunner.startTransaction();
+    const user = await userRepository.findOne(userId, { relations: ['species'] });
+
+    if (!user)
+      throw new HttpException('Error: 존재하지 않는 사용자입니다.', HttpStatus.BAD_REQUEST);
+
+    const animalContent = translateHumanToAnimal(user, patchPostRequestDto);
+
+    await queryRunner.startTransaction();
 
     try {
       const excludedContentIds = excludedPostContents.map((postContent) => postContent.contentsId);
-      await contentRepository.delete(excludedContentIds);
+      if (excludedContentIds.length) await contentRepository.delete(excludedContentIds);
 
       const contents = await contentRepository.save(contentsInfos);
       const postContents = contents.map((content) => ({
         postId: post.id,
         contentsId: content.id,
       }));
-      postContentRepository.insert(postContents);
+      await postContentRepository.insert(postContents);
 
-      await postRepository.update(id, {
+      await postRepository.update(postId, {
         humanContent: patchPostRequestDto.humanContent,
-        animalContent: patchPostRequestDto.animalContent,
+        animalContent: animalContent,
       });
 
       await queryRunner.commitTransaction();
 
       return true;
     } catch (err) {
+      console.log(err);
       await queryRunner.rollbackTransaction();
       return false;
     } finally {
@@ -168,7 +214,6 @@ export class PostService {
     await queryRunner.connect();
 
     const postRepository = queryRunner.manager.getRepository(Post);
-    const postContentRepository = queryRunner.manager.getRepository(PostContent);
     const contentRepository = queryRunner.manager.getRepository(Content);
     const commentRepository = queryRunner.manager.getRepository(Comment);
 
@@ -181,13 +226,13 @@ export class PostService {
     try {
       const contentsIds = post.postContents.map((postContent) => postContent.contentsId);
 
-      contentRepository.delete(contentsIds);
+      await contentRepository.delete(contentsIds);
 
-      if (post.comments.length) commentRepository.remove(post.comments);
+      if (post.comments.length) await commentRepository.remove(post.comments);
 
-      postRepository.delete(post.id);
+      await postRepository.delete(post.id);
 
-      queryRunner.commitTransaction();
+      await queryRunner.commitTransaction();
 
       return true;
     } catch (err) {
@@ -197,4 +242,18 @@ export class PostService {
       await queryRunner.release();
     }
   }
+}
+
+function translateHumanToAnimal(user: User, dto: CreatePostDto | PatchPostRequestDto) {
+  const { sound } = user.species;
+  const contentArr = [];
+  const soundLength = sound.length;
+  const humanContentLength = dto.humanContent.length;
+
+  Array(Math.ceil(humanContentLength / (soundLength + 2)))
+    .fill(0)
+    .forEach((v) => {
+      contentArr.push(sound + '!');
+    });
+  return contentArr.join(' ').slice(0, 501);
 }
